Added create and delete IOQ test cases.
Group GrpBasicInit has included the creation of IOCQ and IOSQ for both
contiguous and discontiguous memories. Additional tests were added to
delete all the Q's which were created.

+++ Queues/queue.h
@@ -41,7 +41,7 @@ public:
     /// Get the memory which encompasses the Q's contents
     uint8_t const *GetQBuffer();
     /// Get the total size in bytes of the Q's memory contents
-    uint64_t GetQSize() { return (mEntrySize * mNumEntries); }
+    uint32_t GetQSize() { return (mEntrySize * mNumEntries); }
 
     /**
      * Send the entire contents of this Q to the logging endpoint
